Move implementation details to DependencyScanningWorker.cpp and dissolve DependencyScannerImpl.h/cpp.
